pyproject.toml
[tool.poetry]
name = "hiyotanwa"
version = "0.1.0"
description = ""
authors = ["root2116 <root6889@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydub = "^0.25.1"
mypy = "^1.10.0"
pydub-stubs = "^0.25.1.1"
ffmpeg-python = "^0.2.0"
openai = "^1.31.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
simpleaudio = "^1.0.4"
pygame = "^2.5.2"
prompt-toolkit = "^3.0.46"
janome = "^0.5.0"
matplotlib = "^3.9.0"
torch = "^2.3.1"
transformers = "^4.41.2"
fugashi = "^1.3.2"
unidic = "^1.1.0"
unidic-lite = "^1.0.8"
scipy = "^1.13.1"
filterpy = "^1.4.5"
scikit-learn = "^1.5.1"
instaloader = "^4.13.1"
umap-learn = "^0.5.6"
hdbscan = "^0.8.39"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
